extract post composer textarea
