+++ apps/web/src/components/Posts/PostComposer.tsx
@@ -1,21 +1,9 @@
-import {
-  autoUpdate,
-  flip,
-  inline,
-  shift,
-  useFloating,
-  useId,
-  useInteractions,
-  useRole,
-} from '@floating-ui/react';
-import { AnimatePresence } from 'framer-motion';
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { useCallback, useState } from 'react';
 import { graphql, useFragment, useLazyLoadQuery } from 'react-relay';
 import styled from 'styled-components';
 
 import ErrorText from '~/components/core/Text/ErrorText';
 import { useModalActions } from '~/contexts/modal/ModalContext';
-import { usePostComposerContext } from '~/contexts/postComposer/PostComposerContext';
 import { useToastActions } from '~/contexts/toast/ToastContext';
 import { PostComposerQuery } from '~/generated/PostComposerQuery.graphql';
 import { PostComposerTokenFragment$key } from '~/generated/PostComposerTokenFragment.graphql';
@@ -25,7 +13,7 @@ import { ChevronLeftIcon } from '~/icons/ChevronLeftIcon';
 import { contexts } from '~/shared/analytics/constants';
 import { GalleryElementTrackingProps, useTrack } from '~/shared/contexts/AnalyticsContext';
 import { useReportError } from '~/shared/contexts/ErrorReportingContext';
-import { MentionType, useMentionableMessage } from '~/shared/hooks/useMentionableMessage';
+import { useMentionableMessage } from '~/shared/hooks/useMentionableMessage';
 import colors from '~/shared/theme/colors';
 
 import breakpoints from '../core/breakpoints';
@@ -33,18 +21,14 @@ import { Button } from '../core/Button/Button';
 import IconContainer from '../core/IconContainer';
 import { HStack, VStack } from '../core/Spacer/Stack';
 import { TitleS } from '../core/Text/Text';
-import { AutoResizingTextAreaWithCharCount } from '../core/TextArea/TextArea';
-import { FloatingCard } from '../Mention/FloatingCard';
-import { MentionModal } from '../Mention/MentionModal';
 import PostComposerNft from './PostComposerNft';
+import { DESCRIPTION_MAX_LENGTH, PostComposerTextArea } from './PostComposerTextArea';
 
 type Props = {
   tokenId: string;
   onBackClick?: () => void;
   eventFlow?: GalleryElementTrackingProps['eventFlow'];
 };
-
-const DESCRIPTION_MAX_LENGTH = 600;
 
 export default function PostComposer({ onBackClick, tokenId, eventFlow }: Props) {
   const query = useLazyLoadQuery<PostComposerQuery>(
@@ -74,6 +58,7 @@ export default function PostComposer({ onBackClick, tokenId, eventFlow }: Props)
           id
         }
         ...PostComposerNftFragment
+        ...PostComposerTextAreaFragment
       }
     `,
     query.tokenById
@@ -90,51 +75,7 @@ export default function PostComposer({ onBackClick, tokenId, eventFlow }: Props)
     handleSelectionChange,
   } = useMentionableMessage();
 
-  const { x, y, reference, floating, strategy, context } = useFloating({
-    placement: 'bottom-start',
-    open: isSelectingMentions,
-    middleware: [flip(), shift(), inline()],
-    whileElementsMounted: autoUpdate,
-  });
-
-  const headingId = useId();
-  const role = useRole(context);
-
-  const { getReferenceProps, getFloatingProps } = useInteractions([role]);
-
-  const textareaRef = useRef<HTMLTextAreaElement | null>(null);
-
-  const { caption } = usePostComposerContext();
-
-  // if there is caption, set it as the default message
-  // for the  share to gallery
-  const prevCaption = useRef<string | null>(null);
-
-  useEffect(() => {
-    if (prevCaption.current !== caption) {
-      prevCaption.current = caption;
-      setMessage(caption);
-      if (textareaRef.current) {
-        const length = textareaRef.current.value.length;
-        textareaRef.current.focus();
-        textareaRef.current.setSelectionRange(length, length);
-      }
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [caption]);
-
-  const setRefs = useCallback(
-    (node: HTMLTextAreaElement | null) => {
-      // Ref's `current` property accepts a DOM node
-      if (node) {
-        textareaRef.current = node;
-        reference(node);
-      }
-    },
-    [reference]
-  );
-
-  const descriptionOverLengthLimit = caption.length > DESCRIPTION_MAX_LENGTH;
+  const descriptionOverLengthLimit = message.length > DESCRIPTION_MAX_LENGTH;
 
   const createPost = useCreatePost();
 
@@ -190,43 +131,6 @@ export default function PostComposer({ onBackClick, tokenId, eventFlow }: Props)
     onBackClick?.();
   }, [onBackClick]);
 
-  const inputPlaceholderTokenName = useMemo(() => {
-    if (!token.name || token.name.length > 30) {
-      return 'this item';
-    }
-    return `"${token.name}"`;
-  }, [token.name]);
-
-  const handleDescriptionChange = useCallback(
-    (event: React.ChangeEvent<HTMLTextAreaElement>) => {
-      setMessage(event.target.value);
-    },
-    [setMessage]
-  );
-
-  const handleSelectMention = useCallback(
-    (item: MentionType) => {
-      selectMention(item);
-
-      const textarea = textareaRef.current;
-      if (textarea) {
-        textarea.focus();
-      }
-    },
-    [selectMention]
-  );
-
-  const handleOnSelect = useCallback(
-    (event: React.MouseEvent<HTMLTextAreaElement, MouseEvent>) => {
-      const target = event.target as HTMLTextAreaElement;
-      const start = target.selectionStart;
-      const end = target.selectionEnd;
-
-      handleSelectionChange({ start, end });
-    },
-    [handleSelectionChange]
-  );
-
   return (
     <StyledPostComposer grow justify="space-between">
       <VStack gap={24}>
@@ -243,39 +147,15 @@ export default function PostComposer({ onBackClick, tokenId, eventFlow }: Props)
         </StyledHeader>
         <ContentContainer>
           <PostComposerNft tokenRef={token} />
-
-          <VStack grow>
-            <AutoResizingTextAreaWithCharCount
-              defaultValue={caption}
-              placeholder={`Say something about ${inputPlaceholderTokenName}`}
-              currentCharCount={message.length}
-              maxCharCount={DESCRIPTION_MAX_LENGTH}
-              textAreaHeight="117px"
-              onChange={handleDescriptionChange}
-              autoFocus
-              hasPadding
-              value={message}
-              ref={setRefs}
-              onSelect={handleOnSelect}
-              {...getReferenceProps()}
-            />
-
-            <AnimatePresence>
-              {isSelectingMentions && (
-                <FloatingCard
-                  className="Popover"
-                  headingId={headingId}
-                  floatingRef={floating}
-                  strategy={strategy}
-                  x={x}
-                  y={y}
-                  getFloatingProps={getFloatingProps}
-                >
-                  <MentionModal keyword={aliasKeyword} onSelectMention={handleSelectMention} />
-                </FloatingCard>
-              )}
-            </AnimatePresence>
-          </VStack>
+          <PostComposerTextArea
+            tokenRef={token}
+            isSelectingMentions={isSelectingMentions}
+            aliasKeyword={aliasKeyword}
+            selectMention={selectMention}
+            setMessage={setMessage}
+            message={message}
+            handleSelectionChange={handleSelectionChange}
+          />
         </ContentContainer>
       </VStack>
       <StyledHStack justify={generalError ? 'space-between' : 'flex-end'} align="flex-end">
